refactor: 평균 별점 API 수정을 대응한다
[#466]
