+++ frontend/src/hooks/tanstack-query/station-details/reviews/useReviewRatings.ts
@@ -4,12 +4,14 @@ import { serverStore } from '@stores/config/serverStore';
 
 import { SERVERS } from '@constants';
 
+import type { StationRatings } from '@type';
+
 export const fetchReviewRatings = async (stationId: string) => {
   const mode = serverStore.getState();
-  return fetch(`${SERVERS[mode]}/stations/${stationId}/total-ratings`).then<number>(
-    async (response) => {
+  return fetch(`${SERVERS[mode]}/stations/${stationId}/total-ratings`).then<StationRatings>(
+    async (response): Promise<StationRatings> => {
       const data = await response.json();
-      return data.totalRatings as number;
+      return data;
     }
   );
 };
